id Data type Int -> Long 으로 변경, 홈 화면 공지 카드 눌렀을 때 공지화면으로 넘어가도록 ClickListener 구현
int64 는 Long

+++ data/src/main/kotlin/us/wedemy/eggeum/android/data/datasource/notice/NoticeDataSourceImpl.kt
@@ -21,7 +21,7 @@ import us.wedemy.eggeum.android.data.util.safeRequest
 public class NoticeDataSourceImpl @Inject constructor(
   private val service: NoticeService,
 ) : NoticeDataSource {
-  override suspend fun getNotice(noticeId: Int): NoticeResponse? {
+  override suspend fun getNotice(noticeId: Long): NoticeResponse? {
     return safeRequest {
       service.getNotice(noticeId)
     }
