+++ data/src/main/kotlin/us/wedemy/eggeum/android/data/repository/NoticeRepositoryImpl.kt
@@ -20,7 +20,7 @@ import us.wedemy.eggeum.android.domain.repository.NoticeRepository
 public class NoticeRepositoryImpl @Inject constructor(
   private val dataSource: NoticeDataSource,
 ) : NoticeRepository {
-  override suspend fun getNotice(noticeId: Int): NoticeEntity? {
+  override suspend fun getNotice(noticeId: Long): NoticeEntity? {
     return dataSource.getNotice(noticeId)?.toEntity()
   }
